feat(cash): 添加备注功能并优化相关操作
- 在 Cash 结构体中添加 note 字段，用于存储备注信息
- 实现 set_note 和 note 方法，用于设置和获取备注信息
- 更新 API.md 文档，添加备注功能的相关说明
- 引入 simple_logger 依赖，用于日志记录
- 修改 init 函数，初始化时设置日志记录器
- 更新 stats 模块，将总计收入和支出的数据类型改为 i64

+++ src/stats.rs
@@ -6,8 +6,8 @@ use anyhow::Result;
 #[derive(serde::Serialize, Debug)]
 pub struct DashboardStats {
     pub total_students: usize,
-    pub total_revenue: i32,
-    pub total_expense: i32,
+    pub total_revenue: i64,
+    pub total_expense: i64,
     pub average_score: f64,
     pub max_score: f64,
     pub active_courses: usize,
